Add TextViewWithPlaceholder class
UIKIt's UITextView class does not provide any support for placeholder
text. There are various ways to add support for this. The approach
chosen here is to subclass UITextView and then:

1. add properties necessary for storing information relevant to the
   placeholder; and

2. register the class with Foundation's default NotificationCenter to
   receive notifications for editing events.

The decision to use NotificationCenter rather than a delegate is to
leave the capability for other classes to register as delegates (which
is what happens in this project's ViewController).

The TextViewWithPlaceholder class is designed to be usable both within
Interface Builder and programmatically. For this reason, at
initialisation, TextViewWithPlaceholder will use the text and colours
to set the placeholder (since this information is displayed in
Interface Builder).