Add example JavaScript processor
Telekinesis is a text manipulation app that runs text inputted by the
user through a 'processor'. A processor is a JavaScript function that is
executed in a JSContext. The JavaScript function is specified in a
JavaScript file that makes a function, process(text), available in the
global namespace.

This commit adds an example of a JavaScript processor to the directory
Processors/. The main View Controller loads the processor and uses it to
process text. The loading does minimal error checking and little
feedback is provided to the user if the JavaScript file has errors in
it.