Use directory for processors rather than group
Xcode has the concepts of both 'directories' and 'groups'. A 'group' is
a logical collection of files that are flattened when the project is
built. The JavaScript processor files were incorrectly saved to the
project as a group when they should have been a directory. This commit
fixes that mistake.