Support pages and section setting item types
Telekinesis lists three different types of items in its Settings screen:
(1) processors; (2) pages; and (3) sections. This commit adds support
for all three.

For pages, it adds the Down library to transform Markdown-formatted
strings into NSAttributedStrings that can be displayed on a single page.
This is intended to be used for pages in a user guide explaining how
processors can be added to Telekinesis.