Detect processor options
The initial design for processors required only that the JavaScript file
exposed a function, process(), that took a single argument as the text
to process. This design did not provide any user-configurable options.
Allowing processors to include options provides a way to avoid multiple
processors that do not differ in terms of their core logic (e.g. a
processor that wraps at 72 characters and a different processor that
wraps at 80 characters).

This commit adds the ability to detect whether a processor has
user-configurable options. This is accomplished by checking whether the
processor has a line that begins with `var process = function(text)`.
Although this heuristic requires the processors to be written in a
certain way, it simplifies the method to check whether the process
function takes user-configurable options.