Add screen for processor options
This commit adds a scene to the main storyboard to display a settings
screen for processors that have options. The screen includes a table of
custom cells that display the name of the option, a text field to enter
a custom value and a comment describing the nature of the option.

This commit also includes changes to the ProcessorModel class that
extract the data from the process() function. JavaScript's reflection
functions are limited so this extraction is achieved by converting the
function definition to a string that is then parsed. This parsing logic
is brittle and may contain bugs.