Use lazy stored properties for processor function and options
Swift provides the `lazy` keyword to mark a stored property as one that
is to be initialised the first time it is used. This is ideal for a
processor's function and options properties as these require the
evaluation of the code before they can be computed.

The previous approach, while avoiding evaluation on initialisation of a
`ProcessorModel object, would evaluate the JavaScript each time the
`function()` method was called. Using the `lazy` keyword avoids this
duplication.