Add persistence to processor options
This commit makes processor options persistent. In adding this
functionality, it became apparent that it would be simpler for the
`ProcessorModel` object to be a reference type so that changes within
one view were automatically reflected across other views.

The changes revealed a bug that caused some option values to overwrite
any subsequent changes. In fixing this bug, the architecture of the app
was revised. In particular, the settings struct was revised to separate
the list of processors from the other settings.