Add processor manager to settings
There are at least 3 vectors along which a user will expect to be able
to adjust the processors in Telekinesis:

(1) enabling and disabling;
(2) reordering;
(3) adding and removing.

This commit adds a processor manager screen to the settings screen that
provides an interface for performing these adjustments. In addition, it
includes an implementation for the first two kinds of adjustment.

As part of adding the screen and providing this implementation, a number
of other parts of the app were refactored. This included the preferences
manager and the settings model.