Support adding processors
A key feature of Telekinesis is the ability for the user to add his or
her own processors. This commit adds rudimentary support for adding
JavaScript files that the user can access. The file is copied to the
sandboxed documents directory where it is recognised on app startup.
Notably, processors are not immediately accessible when added.