Support simultaneous reordering and slide actions
iOS 11 introduced simplified support in UITableView objects for slide
actions.  Unfortunately, these cannot be activated at the same time as
the table is in editing mode.

This commit solves that problems by  using iOS 11's built-in drag and
drop functionality to allow for the reordering of the enabled processors
even when the table is not in editing mode. While the delay before the
drag action is not directly customisable, the amount is reduced by
iterating through the recognisers attach to the table view (this avoid
calling a private API directly).