Store processor scripts under a UUID
Using the base name of the imported JavaScript file as the base name of
the file saved to Telekinesis's documents directory is problematic. Two
files cannot have the same name which may surprise a user who imports
two files with the same name from different locations.

The simplest solution is to use a UUID for the base name and then save
the base name of the imported file in Telekinesis's preferences. This
name can of course be changed at a later point if desired.