Adjust editor and preview heights on keyboard appearance
Previous to this commit, when the keyboard appeared, the bottom inset of
the editor would be set to the height of the keyboard but no other
changes would be made. This resulted in the vertical space for the
editor being reduced while the space for the preview remained unchanged.

This commit instead sets the inset to be half the height of the keyboard
and then reduces the height of the screen by half the height. This is
then split between the editor and the preview.