Use segmented control for text interaction buttons
This commit switches the text interaction buttons from regular buttons
to a segmented control. While one could argue that this violates a
user's expectation, the view has been taken that these are more
obviously buttons with which the user can interact.

The commit also adds a paste button in addition to the reset and copy
controls that were present in the previous design.