Store extension selected processor separately
A user would commonly expect that the processor they select in the app
extension will persist across different invocations of the extension.
In addition, it may be the case that a user wishes to have different
default processors between the app and the extension.

This commit modifies the way the extension loads the selected processor
to instead use the `UserDefaults.standard` object for retrieval and
persistence.