Add heuristic to select an element when using extension with webpages
When using Telekinesis with a webpage, it will first attempt to get
the element with focus. However, on an iPhone, a user will have to blur
the text area being edited in order to open the share menu.

This commit adds a simple heuristic to find the 'intended' text area.
The heuristic simply looks for the text area with the longest value. If
no text area has a value, the heuristic selects the first text area.