Change strategy for displaying large amounts of text
Telekinesis currently uses attributed strings that are generated from
Markdown-formatted text. This approach has the advantage of speed but
is difficult to style.

This commit changes the approach. The About screen is rendered as a
bespoke view while all the other screens of large amounts of text are
rendered by WKWebView. The text for these screen is still Markdown but
rather than being stored in the dictionary of localised strings, the
pages are kept as separate Markdown files which are assembled at
runtime.