Fix safe area adjusting in the action extension on iPads
Action extensions are displayed differently on iPads than on other
iOS devices. On iPads, the extension floats in a modal window that is
centred on the screen. The modal is shifted up when the keyboard
appears. The result of this is that it's possible for the keyboard never
to cover the extension's view or to only cover it partially. The problem
is that the code used in the extension for adjusting the height of the
safe area assumes that the entire extension is covered by the keyboard.

This commit includes a fix that is, frankly, a hack. The adjusting
function now checks if the device is an iPad. If it is, rather than
simply adjusting the safe area to include the height of the keyboard,
the safe area is adjusted to include the overlapping portion of the
keyboard. Or that's the intention. It doesn't work properly for reasons
that could not be determined and instead an additional 50 points
needed to be added to the adjusted height.