Add preview hiding for portrait orientation
Scott Yoshinaga suggested that the preview may be something that users
would like to hide. This is relatively straightforward to do when the
app is being used on an iPhone in portrait orientation. It is a more
complicated problem for the other devices and orientations. Given the
larger amount of screen space in those situations, this commit adds the
feature only for portrait orientation.