Use editor view controller for action extension view controller
The initial implementation of the action extension view controller used
a copy and paste of the editor view controller used in the main app.
This was helpful for initial development but creates duplication and the
difficulty of keeping the features of the two editors in sync. It also
led to differing approaches for the layout of the editors in the
respective storyboards.

This commit refactors the action extension view controller to subclass
the editor view controller. The action extension storyboard is updated
to work with these changes. In addition to reducing duplication, this
has the added benefited of solving a UI bug with the on-screen keyboard
and the action extension's editor in landscape orientation.