Synchronise scroll position between preview and editor
It would be a nicer user experience if the text preview scrolled in
sync with the text editor. However, given that the number of characters
in the preview will often not match the number of characters in the
editor, there is no simple way to do this.

This commit uses a heuristic that takes the ratio of the position of the
cursor in the text editor to the number of characters in the text editor
and calculates a position with an equivalent ratio in the text preview.
If the cursor is at the beginning or end of the editor, the heuristic
is not used.