Fix option updating
At some point, an additional view had been added between the text field
in the Options section and the table cell. That meant that the method
updating the option value in the relevant processor broke.

This fixes the problem by adding an additional `superview()` call but
this is clearly a brittle solution and a better one should be found.