Move view objects into separate files
There are a number of views which were defined in the view controller
where they were used. While this approach makes sense for views which
are specific to the layout of the view controller's scene, it does not
make sense for general use views. This commit extracts those general
use views into separate files.