Add separate Mac version
This commit adds Mac support via a separate target. The app is still
built with Mac Catalyst but now uses a separate storyboard that is more
appropriate for a Mac version.

This commit also introduces the concept of 'favourites'. Unlike in the
existing target, the Mac version does not delineate between enabled and
disabled processors. Rather, certain processors can be marked as
'favourited'. These favourited processors are listed twice: once in
either the built-in or user-added sections and once in the favourited
section.