Add workaround for dynamic height footer bug
When adding a label to the footer of a table view, the footer will not
resize properly if the text size is increased. The reason seems to be a
bug in the rendering of dynamic height footers. This commit adds code to
programmatically calculate and display the footer with the correct
height.