Make Mac Catalyst version part of Flext target
The Mac Catalyst version was developed as a separate target but that
doesn't seem like the best long-term approach. This commit moves the
files used by the Mac Catalyst version into the main Flext target. It
loads a separate storyboard for the main interface by defining the
storyboard to use a variable name `${MAIN_STORYBOARD}` and then sets
that conditionally depending on the SDK (either `Main` or `MainMac`).